scrapper/src/main/java/edu/java/clients/GitHubClient.java
package edu.java.clients;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import org.springframework.web.reactive.function.client.WebClient;
import reactor.core.publisher.Mono;

@Component
public class GitHubClient {
    private final WebClient webClient;

    @Autowired
    public GitHubClient(WebClient.Builder webClientBuilder) {
        this.webClient = webClientBuilder.baseUrl("https://api.github.com").build();
    }

    @Autowired
    public GitHubClient(WebClient.Builder webClientBuilder, String baseUrl) {
        this.webClient = webClientBuilder.baseUrl(baseUrl).build();
    }

    public Mono<String> fetch(String owner, String repoName) {
        return webClient.get()
            .uri("/repos/{owner}/{repo}", owner, repoName)
            .retrieve()
            .bodyToMono(String.class);
    }
}
